fix: headless imports
