+++ apps/main-app/ts/features/itwallet/statusList/saga/index.ts
@@ -7,7 +7,7 @@ import { itwLifecycleIsValidSelector } from "../../lifecycle/store/selectors";
 import {
   registerItwStatusListFetchTask,
   unregisterItwStatusListFetchTask
-} from "../tasks";
+} from "../tasks/manager";
 
 /**
  * Registers the ITW Status List fetch task with expo-background-task.
