+++ apps/main-app/ts/features/itwallet/statusList/tasks/index.ts
@@ -1,9 +1,5 @@
 import * as BackgroundTask from "expo-background-task";
 import * as TaskManager from "expo-task-manager";
-import {
-  trackItwStatusListFetchRegistered,
-  trackItwStatusListFetchRegisterFailure
-} from "../analytics";
 import { storeLastStatusListCheckTimestamp } from "../utils/storage";
 
 /**
@@ -18,50 +14,6 @@ export const ITW_STATUS_LIST_FETCH_TASK = "io-itw-status-list-fetch";
  * Note that the actual execution timing is determined by the OS and may vary.
  */
 export const ITW_STATUS_LIST_FETCH_TASK_INTERVAL_MINUTES = 60 * 4;
-
-/**
- * Registers the ITW Status List background fetch task with expo-background-task
- * if the background task API is available and the task is not already registered.
- */
-export const registerItwStatusListFetchTask = async (): Promise<void> => {
-  try {
-    const isRegistered = await TaskManager.isTaskRegisteredAsync(
-      ITW_STATUS_LIST_FETCH_TASK
-    );
-    if (isRegistered) {
-      return;
-    }
-
-    await BackgroundTask.registerTaskAsync(ITW_STATUS_LIST_FETCH_TASK, {
-      minimumInterval: ITW_STATUS_LIST_FETCH_TASK_INTERVAL_MINUTES
-    });
-    trackItwStatusListFetchRegistered();
-  } catch (error) {
-    trackItwStatusListFetchRegisterFailure(
-      error instanceof Error ? error.message : "unknown"
-    );
-  }
-};
-
-/**
- * Unregister the ITW Status List background fetch task with expo-background-task.
- *
- * @throws Will throw an error if the unregistration fails
- */
-export const unregisterItwStatusListFetchTask = async (): Promise<void> => {
-  try {
-    const isRegistered = await TaskManager.isTaskRegisteredAsync(
-      ITW_STATUS_LIST_FETCH_TASK
-    );
-    if (!isRegistered) {
-      return;
-    }
-
-    await BackgroundTask.unregisterTaskAsync(ITW_STATUS_LIST_FETCH_TASK);
-  } catch {
-    // Ignore errors during unregistration
-  }
-};
 
 /**
  * Register the ITW Status List fetch task handler with expo-task-manager.
