+++ apps/main-app/ts/features/itwallet/statusList/tasks/manager.ts
@@ -1,0 +1,59 @@
+import * as BackgroundTask from "expo-background-task";
+import * as TaskManager from "expo-task-manager";
+import {
+  trackItwStatusListFetchRegistered,
+  trackItwStatusListFetchRegisterFailure
+} from "../analytics";
+import {
+  ITW_STATUS_LIST_FETCH_TASK,
+  ITW_STATUS_LIST_FETCH_TASK_INTERVAL_MINUTES
+} from ".";
+
+export {
+  ITW_STATUS_LIST_FETCH_TASK,
+  ITW_STATUS_LIST_FETCH_TASK_INTERVAL_MINUTES
+};
+
+/**
+ * Registers the ITW Status List background fetch task with expo-background-task
+ * if the background task API is available and the task is not already registered.
+ */
+export const registerItwStatusListFetchTask = async (): Promise<void> => {
+  try {
+    const isRegistered = await TaskManager.isTaskRegisteredAsync(
+      ITW_STATUS_LIST_FETCH_TASK
+    );
+    if (isRegistered) {
+      return;
+    }
+
+    await BackgroundTask.registerTaskAsync(ITW_STATUS_LIST_FETCH_TASK, {
+      minimumInterval: ITW_STATUS_LIST_FETCH_TASK_INTERVAL_MINUTES
+    });
+    trackItwStatusListFetchRegistered();
+  } catch (error) {
+    trackItwStatusListFetchRegisterFailure(
+      error instanceof Error ? error.message : "unknown"
+    );
+  }
+};
+
+/**
+ * Unregister the ITW Status List background fetch task with expo-background-task.
+ *
+ * @throws Will throw an error if the unregistration fails
+ */
+export const unregisterItwStatusListFetchTask = async (): Promise<void> => {
+  try {
+    const isRegistered = await TaskManager.isTaskRegisteredAsync(
+      ITW_STATUS_LIST_FETCH_TASK
+    );
+    if (!isRegistered) {
+      return;
+    }
+
+    await BackgroundTask.unregisterTaskAsync(ITW_STATUS_LIST_FETCH_TASK);
+  } catch {
+    // Ignore errors during unregistration
+  }
+};
